Remove `include/grpcpp/impl/codegen/security/auth_context.h`
